feat:开发提交订单接口

+++ mall-order/mall-order-service/src/main/java/com/salute/mall/order/service/pojo/entity/SaleOrderDetail.java
@@ -53,7 +53,7 @@ public class SaleOrderDetail implements Serializable {
 
     @ApiModelProperty(value = "购买数量",name = "buyQty")
     @TableField("buy_qty")
-    private BigDecimal buyQty;
+    private Integer buyQty;
 
     @ApiModelProperty(value = "商品主图",name = "mainImage")
     @TableField("main_image")
